chore(setup): initialize phoenix application
The phoenix application has been generated using the `phoenix.new` mix
task. The `no-brunch` switch was passed to this task as it is not
anticipated that we will be using the asset pipeline for this project.

The static files have been removed as they are not required for an
endpoint api.

The config has been updated to check for the existence of an environment
secret file `<env>.secret.exs` rather than just have a default
`prod.secret.exs`. The `*.secret.exs` files have also been added to
`.gitignore`