feat(User): add user model with CRUD routes
A User model has been added that holds the basic account credentials of
a user.

The route identifier for each endpoint is the `username` and not the
`id` as this will prevent guessing user accounts by incrementing the id
manually.

There is no authentication for the endpoints at the moment, but this
will be implemented in the future.

The `Comeonin` dependency has been included to enable password hashing
for security.