feat(AccessToken): add authentication via tokens for certain endpoints
A number of RESTful endpoints require authentication as they perform
sensitive operations. To provide the authentication a token based system
has been implemented.

A new `/auth` route can be POSTed to with user credentials which will
return an `access_token` with a limited lifespan. This `access_token`
can be passed in the `authorization: Bearer` HTTP header when making
requests to protected endpoints.

If the authorization is successful then the endpoint will operate as
normal, however if the authorization fails then an appropriate 401 HTTP
status code is returned.

A collection of `/auth` routes now are present for the following
actions:

  * `POST /auth` -> Creates token for user (i.e. login)
  * `GET /auth/:username` -> Retrieves existing token information
  * `DELETE /auth/:username` -> Deletes token for user (i.e. logout)

Error messages returns have also been formalised to the `jsonapi`
standard. Now all error messages are returns with the appropriate HTTP
status code in the header and the body in the format of:

```
{
  errors: { title: "<error_message>" }
}
```

Some of the previous `user_controller` tests have been removed as they
are reliant on authorized endpoints and therefore the errors could never
be raised.

E.g. - Deleting a user with an invalid username raises a 404. This could
never happen because the delete user action would never execute if the
user did not exist as the authorization would fail and return earlier.