feat(Status): add status model with CRUD routes
A Status model has been added that holds the different status
descriptions for a user. It also flags which status is the active one.

There is no `GET / :show` route as it is not relevant to display a
single Status, it is replaced with a `GET / :list` route that displays
all of the user's statuses.

All Status endpoints require authentication before accessing.

It's also worth noting that a Status cannot be created with an active
Status of true as this would cause issues by users accidentally setting
their Status when creating a new one, which is undesireable.

When a Status is updated and it is set to active:true then the previous
active:true Status is automatically set to false, to ensure only one
Status can ever be active:true.

It is also not possible to update the active:true Status to
active:false, which means the only way to set a Status to active:false
is by setting another Status to active:true. This ensures that at least
one Status has to be active:true.

It is also not possible to delete the active:true Status, this means
there must always be at least one Status that is active and means the
user can never be left without a Status.