chore(authentication): remove username from url in authenticated routes
Previously, the username was required as a url parameter when calling
any endpoint that needed authentication. This was because the token
lookup also required to know about the username as added protection.

However, it's better to just generate a token using the username as a
seed to prevent potential `access_token` duplication. This means that
the authentication relies only on the `access_token`.

Also, the `/users` route has been renamed to `/user` as it only ever
deals with an indivual user and therefore is confusing using the plural
to identify the resource.