feat(User): require old password on password update
Previously, the `PUT /user` route allowed the password to be changed by
simply providing a `password` POST parameter. This meant that the
user's password could be updated by anyone who had a valid
`access_token` meaning it could be easier to compromise a user's
account.

Now, when updating a user's password via the `PUT /user` route along
with a valid `access_token` the existing password must also be provided
in the format of:

```
{
  "old_password": <existing_password>,
  "new_password": <new_password>
}
```

If the `old_password` is invalid then a `401 Unauthorized` will be
returned.

It is no longer possible to pass a `password` parameter via the
`PUT /user` route as it gets stripped from the parameters server side.

Also, when a user's password is successfully changed all previously
created access tokens for that user are deleted ensuring that a new
`access_token` is required using the newly created password.

The queries on the `access_token` model have also been renamed to all
specify `get` in the function name as they are just generic queries and
not related to the action they may perform.