fix(schema): update schema queries to not contain action verb
Previously, schemas for the models had defined queries that contained
verbs about which action the query was to perform, e.g. get_by_id or
update_by_name. This is not correct as the queries can be used to select
items for any action, therefore they are purely referenced now by the
fields the query will be performed against.