feat(User): create default status for new user
When a new user is created, a default active status is set to ensure
integrity in the user status data (because a user always needs 1 active
status). As there is a default active status, the user can only create
new statuses and set another one to be active before deleting this
initial status. Therefore, the integrity model of the user always having
at least 1 active status will hold up.