feat(Follow): add follow model with CRD routes
A Follow model has been added that allows a User to Follow another user
and this relation then determines a user's follower/following list.

The follower/following list can be viewed for the current authenticated
user by calling `GET /follow` and this will always return the user's
follower/following list.

The follower/following list can be viewed for any other user on an
unauthenticated route at `GET /follow/:username` however in the future
this route may return an error if/when user access restrictions are
implemented.