fix(authentication): prevent crash when logging in without credentials
If a log in attempt was made without providing the correct form
parameters (`username` and `password`) the service would crash because
it could not match the correct function to call.

This has now been guarded against and returns a `401` status code if the
relevant fields are not present as authentication cannot happen.