feat(authentication): add refresh_token to gain authentication
Previously, the only way to authenticate and receive an `access_token`
was to log in with a username and password. However, because the
`access_token` has a short lifespan, it would mean that the user would
have to re-authenticate with their username and password when the
`access_token` had expired.

A better solution is to allow a `refresh_token` to be returned when a
user authenticates and this `refresh_token` can be used to authenticate
the user (without the need for providing the username and password).

The authentication request now accepts the following scenarios:

  * username/password authentication

```
{
  grant_type: "password",
  username: <username>,
  password: <password>,
  client_id: <client_id>
}
```

Note: the `client_id` is optional in username/password authentication,
but if it is not provided then a `refresh_token` will not be generated.

  * refresh_token authentication

```
{
  grant_type: "refresh_token",
  refresh_token: <refresh_token>,
  client_id: <client_id>
}
```

Either of the authentication methods above will return the same
response, which a `201 Created` and the payload is the same
authentication payload as before but it includes the `refresh_token`,
however the `refresh_token` may be blank if one was not created (if a
`client_id` was not provided).

At present, the `client_id` is just a string provided by the client to
the service. It can be anything providing it is over 8 characters long
and the purpose of it is to ensure the `access_token` refresh is coming
from the client that the user originally logged in with.

In the future, the `client_id` may be a pre-determined list of allowed
clients that is created and stored in the database service side to limit
which clients are allowed to use the service.