feat(User): ensure user query by username is case insensitive
Previously, any request for a User by the username had to have the
parameter provided in the correct case. For example, when logging in the
username had to be precisely the correct case. It is preferable to be
able to provide the username in any case and it still match.

To ensure a duplicate user cannot be created by simply providing a
username in a different case a new `display-name` field has been added
to the user schema. The display name is returned in the correct case
when querying a user as the `username` field is now lowercased in the
database so the unique check is always case insensitive.