feat(Search): add ability to list statues by user or status
It is now possible to receive a list of statuses that match a given
status query or list users by user details and their current status.

`search/status/:query`

Will perform a status search for the given query and return all users
with an active status that is like the given query.

`search/user/:query`

Will perform a user search for the given query and return all users and
their current active status for any user that is like the given query.

Optional query string parameters can be provided to add pagination to
the search results:

`search/user/:query?limit=10&offset=5`

Will perform a search and limit the results to 10 records starting at
record 5.