chore(Search): move search queries to history model
As the search queries actually use the `History` table for their
results it makes more sense that the query function is inside the
`History` model.

Also, to prepare for a future addition, the `list_by` function now
defines the type as `model` because it is actually used to specify a
model name.

This means that `type` is now used to specify the query type. At
present `:default` is the only type which simply means there is no
extra sorting or filtering applied.