feat(List): add routes to list top n resources
The ability to call a new `GET list/` route to provide a `top n`
function to show statistics.

The following new routes are available:

  * `/list/new/users` - returns a list of users with newest first
  * `/list/new/statuses` - returns a list of users with latest active
  * `/list/popular/statuses` - returns a list of highest active status

The lists can be paginated by passing through `limit` and `offset` query
parameters like:

`/list/new/users?limit=10&offset=5`

An update to the `TestCommon.create_user` function now allows for the
user `inserted_at` time to be modified by providing an additional
parameter which shifts the `inserted_at` and `updated_at` time for a
newly created user by the number of seconds given in the parameter.

This is required for ensuring the tests for the new `/list` routes are
consistent.

The way in which limit and offset were being retrieved from the params
has now been formalized in a new module `Util.params` and has a function
called `get_limit_offset/1` which takes the params and returns a tuple
of the limit and offset variables.

The `History.list_by` function has had the type updated from `:default`
to `:search` as it's more descriptive of the type of query being done.