fix(List): rework list retrieval to use status model
The History model is not appropriate for performing queries for active
information like `newest active status` or `newest users` because it
takes into account all previous statuses for the user meaning
duplicates are returned. This was missed in the tests but now the tests
have been improved to highlight the issue with the old way or working.

The Status model is the more appropriate place to perform the queries as
it works on active data due to the `active` boolean flag on the record.