fix(Search): rework search to use status model
As with the List function, the `History` model is not the best place to
query search results as it can return duplicates. The `Status` model
relates to active data so it makes more sense as the starting point for
dealing with search queries that require the active status to be
returned.

The tests have also been updated to highlight the duplicates issue with
the previous code.