feat(deploy): add exrm dependency to manage deployment
Exrm has now been included as a dependency to manage the creation of a
release binary for deployment.

The `:connection` app has been added to the applications list as it
seems that `Ecto` or `Postgrex` doesn't add this sub dependency and
would cause the exrm deployment to crash.

The `corsica` dependency has also been added to the applications list as
it was previously omitted and is required.

The `config/prod.exs` has been updated to set the Endpoint to act as a
server and the `cache_static_manifest` has been removed as there are no
static files compiled for the api.

The `/rel` directory created by exrm has also been added to the
`.gitignore` file as it should not be commited to the repo.